requirements.txt
# Core
numpy>=1.21.0
torch>=1.10.0
matplotlib>=3.5.0
networkx>=2.6.0

# RL/ML
gymnasium>=0.28.1
gym>=0.26.2

# Utils
tqdm>=4.62.0
pyyaml>=6.0
pandas>=1.3.0
arabic-reshaper>=3.0.0
python-bidi>=0.6.7

# Development
pytest>=6.2.5
black>=21.12b0
flake8>=4.0.0

# To install all dependencies: pip install -r requirements.txt
